[Player] 네트워킹 APIManger 로 변경
