+++ FLO/Network/APIManager.swift
@@ -7,7 +7,7 @@
 
 import Foundation
 
-class HttpClient<T: Codable> {
+class APIManger<T: Codable> {
     
     let baseUrl = "https://grepp-programmers-challenges.s3.ap-northeast-2.amazonaws.com/2020-flo/song.json"
     
@@ -41,5 +41,23 @@ class HttpClient<T: Codable> {
             completed(Result.success(data))
         }
         dataTask.resume()
+    }
+    
+    /// JSON Data 를 Music 모델로 파싱해서 반환하는 함수
+    func getMusic(completed: @escaping (Music) -> Void) {
+        
+        let urlRequest = self.getUrlRequest(self.baseUrl)
+        
+        self.getJsonData(of: urlRequest) { (result) in
+            if let jsonData = try? result.get() {
+                do {
+                    let decoder = JSONDecoder()
+                    let response = try decoder.decode(Music.self, from: jsonData)
+                    completed(response)
+                } catch {
+                    completed(Music.EMPTY)
+                }
+            }
+        }
     }
 }
